+++ handlers/ars9_generic_update_v1/test/simple_input.rb
@@ -1,0 +1,13 @@
+{
+  'info' => {
+    'api_server' => 'http://yourserver:8008/api',
+    'api_username' => '',
+    'api_password' => '',
+    'enable_debug_logging' => 'Yes'
+  },
+  'parameters' => {
+    'error_handling' => "Error Message",
+	'form' => "CTM:Region",
+    'field_values'          => '{"Region": "B"}',
+    'entry_id' => 'RGN000000000201'
+}}