+++ bridge-adapters/kinetic-bridgehub-adapter-ars/changelog.md
@@ -1,3 +1,6 @@
 ARS v1.0.1 (12-23-19)
     * fixed issue with multiple remedy instances using the same bridge adapter.
     Global forms were being set for performance reasons.
+
+ARS v1.0.3 (12-27-19)
+    * Removed nexus from the POM.  We're using s3 to store dependencies now.